Reorganize enum utilities
